requirements.txt
cffi==1.15.0
msgpack==1.0.3
py-algorand-sdk==1.8.0
pycparser==2.21
pycryptodomex==3.12.0
PyNaCl==1.5.0
tinyman-py-sdk==0.0.4
